+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 12.5
 - Fix HTML error label's for attribute doesn't match any element id in General_Info.php
 - Fix add .tinymce CSS class when class already set in $extra param in Inputs.php
 - JS add CSP root, functions & programFunctions (global, always loaded) in Gruntfile.js, assets/js/csp/
+- JS add setInnerHTML() function in warehouse.js
 
 Changes in 12.4.3
 -----------------